Add class for JSON Deserialization and HttpStatusCode check

+++ StockPriceWatcher/Classes/Stock.cs
@@ -3,16 +3,16 @@
     internal class Stock
     {
         private string symbol;
-        private double regularMarketPrice;
-        private double buyPrice;
-        private double sellPrice;
+        private decimal regularMarketPrice;
+        private decimal buyPrice;
+        private decimal sellPrice;
 
         public string Symbol { get => symbol; set => symbol = value; }
-        public double RegularMarketPrice { get => regularMarketPrice; set => regularMarketPrice = value; }
-        public double BuyPrice { get => buyPrice; set => buyPrice = value; }
-        public double SellPrice { get => sellPrice; set => sellPrice = value; }
+        public decimal RegularMarketPrice { get => regularMarketPrice; set => regularMarketPrice = value; }
+        public decimal BuyPrice { get => buyPrice; set => buyPrice = value; }
+        public decimal SellPrice { get => sellPrice; set => sellPrice = value; }
 
-        public Stock(string stockName, double regularMarketPrice, double sellPrice, double buyPrice) 
+        public Stock(string stockName, decimal regularMarketPrice, decimal sellPrice, decimal buyPrice) 
         {
             this.symbol = stockName;
             this.sellPrice = sellPrice;
